placement_platform/requirements.txt
streamlit
google-genai
pypdf2
pandas